create party endpoint

+++ database/postgres.js
@@ -95,8 +95,8 @@ name, host_id, is_free, min_price, max_price,
 location, start_time, end_time, min_rating, type, invite_limit
 ) VALUES('${name}', ${Number.parseInt(hostID)}, ${isFree},
 ${Number.parseInt(minPrice)}, ${Number.parseInt(maxPrice)}, '${address}',
-${Number.parseDouble(start)}, ${Number.parseDouble(end)},
-${Number.parseDouble(minRating)}, ${Number.parseInt(type)},
+${Number.parseFloat(start)}, ${Number.parseFloat(end)},
+${Number.parseFloat(minRating)}, ${Number.parseInt(type)},
 ${Number.parseInt(limit)})`, (err, result) => {
         done();
         if(err) resolve(null);
